Refactor: worker の fake テストダブルを testkit 規約へリネーム
worker/fake（package fake）を worker/testkit（package testkit）へリネームし、
コンストラクタ New を NewFake に変更。同じ boundary 配下の clock/testkit や
repo 共通の testkit 命名規約に揃えた。worker/mock（生成）+ worker/testkit
（手書き）の構成が clock/mock + clock/testkit と並行になる。
