+++ internal/usecase/boundary/worker/testkit/test_kit.go
@@ -1,6 +1,6 @@
-// Package fake は、worker seam（Consumer / FailureHandler）の in-memory テストダブルを提供します。
+// Package testkit は、worker seam（Consumer / FailureHandler）の in-memory テストダブルを提供します。
 // 実 broker 無しで engine の不変条件テストを green にするための 2nd impl です（SDK 非依存）。
-package fake
+package testkit
 
 import (
 	"context"
@@ -42,8 +42,8 @@ type Fake struct {
 	notify      chan struct{} // long-poll 起床用のブロードキャスト
 }
 
-// New は、空の Fake を生成します。
-func New() *Fake {
+// NewFake は、空の Fake を生成します。
+func NewFake() *Fake {
 	return &Fake{
 		inflight:     make(map[string]worker.Message),
 		delivery:     make(map[string]int),
